pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "relief-blend"
version = "0.1.0"
description = "Multiply-blend shaded relief onto matplotlib/earthkit-plots maps"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "numpy",
    "pillow",
    "matplotlib",
]

[project.optional-dependencies]
earthkit = ["earthkit-plots", "cartopy"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
relief_blend = ["data/*.jpg"]
